+++ yarn-project/aztec.js/src/contract/checker.test.ts
@@ -1,4 +1,4 @@
-import { abiChecker } from './parser.js';
+import { abiChecker } from './checker.js';
 
 describe('abiChecker', () => {
   describe('no functions', () => {
